requirements.txt
pytest-playwright==0.3.3
pytest_html==3.2.0